+++ cmd/envNew.go
@@ -19,9 +19,9 @@ end result will be a composite environment.`,
 		if common.DebugFlag {
 			defer common.Stopwatch("New environment creation lasted").Report()
 		}
-		ok := conda.MustConda()
+		ok := conda.MustMicromamba()
 		if !ok {
-			pretty.Exit(2, "Could not get miniconda installed.")
+			pretty.Exit(2, "Could not get micromamba installed.")
 		}
 		label, err := conda.NewEnvironment(forceFlag, args...)
 		if err != nil {