+++ robot/config.go
@@ -229,10 +229,8 @@ func (it *Activity) ExecutionEnvironment(base Robot, location string, inject []s
 	}
 	return append(environment,
 		"CONDA_DEFAULT_ENV=rcc",
-		"CONDA_EXE="+conda.BinConda(),
 		"CONDA_PREFIX="+location,
 		"CONDA_PROMPT_MODIFIER=(rcc)",
-		"CONDA_PYTHON_EXE="+conda.BinPython(),
 		"CONDA_SHLVL=1",
 		"PYTHONHOME=",
 		"PYTHONSTARTUP=",