+++ docs/changelog.md
@@ -1,10 +1,21 @@
 # rcc change log
 
-## v7.1.5
+## v8.0.0 (date: 5.1.2021)
+
+- BREAKING CHANGES
+- removed miniconda3 download and installing
+- removed all conda commands (check, download, and install)
+- environment variables `CONDA_EXE` and `CONDA_PYTHON_EXE` are not available
+  anymore (since we don't have conda installation anymore)
+- adding micromamba download, installation, and usage functionality
+- dropping 32-bit support from windows and linux, this is breaking change,
+  so that is why version series goes up to v8
+
+## v7.1.5 (date: 4.1.2021)
 
 - now command `rcc man changelog` shows changelog.md from build moment
 
-## v7.1.4
+## v7.1.4 (date: 4.1.2021)
 
 - bug fix for background metrics not send when application ends too fast
 - now all telemetry sending happens in background and synchronized at the end